Fundamental Kotlin - Scope functions.
